Didn't work the whole way

+++ deploy-ghpages.sh
@@ -11,8 +11,11 @@
  cp ../app.js ./app.js
  
  git add .
-	 
- if [[ `git status --porcelain` ]]; then
+
+ if output=$(git status --porcelain) && [ -z "$output" ]; then
+   # Working directory clean
+ else 
+   # Uncommitted changes
  	git commit -m "Deployed to Github Pages"
  	git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master:gh-pages > /dev/null 2>&1
  fi
